// apps/core_microservice/src/modules/users/dto/create_user.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Role } from '@prisma/client';
import { IsEnum } from 'class-validator';
export class CreateUserDto {
  @ApiProperty({
    description: 'id is generated by default in prisma',
  })
  id?: string;

  @ApiProperty({
    description:
      'Role is enum and can only have one of two values user or admin',
  })
  @IsEnum(Role)
  Role: Role;

  @ApiProperty({
    description:
      'Disabled is boolean that tells if account is disabled by default is false',
  })
  disabled?: false;
}